requirements.txt
fastapi==0.115.9
uvicorn==0.34.2
openai==1.78.1
chromadb==1.0.9
requests==2.32.3
python-dotenv==1.1.0 